requirements.txt
openai==2.14.0
PyYAML==6.0.3
